refactor: move code from left & top bars to new files
